changed type to decimal for funds

+++ src/openg2p_portal_api/models/program.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import Optional
 
 from pydantic import BaseModel, ConfigDict, Field, validator
@@ -40,8 +41,8 @@ class ProgramSummary(BaseModel):
 
     program_name: Optional[str] = None
     enrollment_status: Optional[str] = None
-    total_funds_awaited: Optional[int] = None
-    total_funds_received: Optional[int] = None
+    total_funds_awaited: Optional[Decimal] = None
+    total_funds_received: Optional[Decimal] = None
 
 
 class ApplicationDetails(BaseModel):
@@ -58,7 +59,7 @@ class BenefitDetails(BaseModel):
 
     program_name: Optional[str] = None
     enrollment_status: Optional[str] = None
-    funds_awaited: Optional[int] = None
-    funds_received: Optional[int] = None
+    funds_awaited: Optional[Decimal] = None
+    funds_received: Optional[Decimal] = None
     entitlement_reference_number: Optional[int] = None
     # cycle_name: Optional[str]=None
